+++ CHANGELOG.md
@@ -12,6 +12,11 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - `clairvoyant.sightToken` command.
 - Added support for text editor context menu.
 - Added support for editor title menu.
+
+### Changed
+
+- When commands cannot be used, it are not displayed on the command palette.
+
 ## 5.0.1 - 2019-12-05
 
 ### Added